validate last state

+++ test-client/src/AutoTest.js
@@ -5,9 +5,54 @@ const NEXT_STATES = ["waiting", "entering", "looping", "on_break", "exiting"];
 
 /*
 Normal run: all chains resolve, promise.all resolves
-User stop: queued tasks check isRunning and resolve early, promise.all resolves
-Any shuttle error: whole run stops. that chain rejects, promise.all rejects, all further errors are ignored
+User stop: queued tasks check isRunning and resolve silently, promise.all resolves
+Any shuttle error: entire test fails. error task throws, promise.all throws, further errors ignored
 */
+export async function executeTest(testData) {
+    if (isRunning) {
+        warnUser("Test already running");
+        return;
+    }
+
+    console.log("Starting automated test");
+    isRunning = true;
+    try {
+        // build all shuttle event chains
+        for (const shuttle of testData.shuttles) {
+            // queue addShuttle first, before the test case events
+            enqueue(shuttle.id, addShuttle);
+            let lastEvtType = null;
+            for (const evt of shuttle.events) {
+                if (!NEXT_STATES.includes(evt.type)) {
+                    throw new Error(`Unexpected event type ${evt.type}`);
+                }
+                const prev = lastEvtType;
+                enqueue(shuttle.id, () => executeEvent(shuttle.id, evt, prev));
+                lastEvtType = evt.type;
+            }
+        }
+    } catch (err) {
+        warnUser(err);
+        isRunning = false;
+        eventChains.clear();
+        return;
+    }
+
+    console.log("Built shuttle event chains", eventChains);
+    try {
+        // concurrently execute all shuttle event chains
+        await Promise.all([...eventChains.values()]);
+        console.log("All event calls have been executed");
+        console.log("Test finished successfully");
+    } catch (err) {
+        warnUser(err);
+        console.log("Test failed");
+    } finally {
+        isRunning = false;
+        eventChains.clear();
+    }
+}
+
 function enqueue(shuttleId, task) {
     if (!isRunning) {
         throw new Error("enqueue() called while test is not running");
@@ -19,7 +64,6 @@ function enqueue(shuttleId, task) {
                 await task();
             } catch (err) {
                 isRunning = false;
-                warnUser(err);
                 throw err;
             }
         }
@@ -27,71 +71,60 @@ function enqueue(shuttleId, task) {
     eventChains.set(shuttleId, next);
 }
 
-export async function executeTest(testData) {
-    if (isRunning) {
-        warnUser("Automated test already running");
-        return;
-    }
-
-    console.log("Starting automated test");
-    isRunning = true;
-    try {
-        // queue all shuttle events
-        for (const shuttle of testData.shuttles) {
-            // add shuttle first, then add its events
-            enqueue(shuttle.id, addShuttle);
-            for (const evt of shuttle.events) {
-                if (!NEXT_STATES.includes(evt.type)) {
-                    throw new Error(`Unexpected event type ${evt.type}`);
-                }
-                enqueue(shuttle.id, () => executeEvent(shuttle.id, evt));
-            }
-        }
-    } catch (err) {
-        warnUser(err);
-        isRunning = false;
-        eventChains.clear();
-        return;
-    }
-
-    console.log("Loaded shuttle event chains", eventChains);
-    try {
-        // concurrently execute all shuttle event chains
-        await Promise.all([...eventChains.values()]);
-        console.log("Automated test finished successfully");
-    } finally {
-        isRunning = false;
-        eventChains.clear();
+async function executeEvent(id, evt, lastEvtType) {
+    await verifyCurrentState(id, lastEvtType);
+    await setNextState(id, evt.type);
+    console.log("set next state for", id, evt);
+    if (evt.type === NEXT_STATES[0] || evt.type === NEXT_STATES[3]) {
+        await new Promise(resolve => setTimeout(resolve, evt.duration * 1000));
     }
 }
 
-async function executeEvent(id, evt) {
-    await validateState(id);
-    console.log("validiated", id);
-    await setState(id, evt.type);
-    console.log("set next state", id, evt);
-}
+// This function waits until shuttle.next_state is free
+// If, at that time, shuttle.state is mismatched, throw an error
+// This ensures correctness (last state must equal last test case event)
+async function verifyCurrentState(id, lastEvtType) {
+    let checkCount = 0, checkLimit = 3;
+    const checkLast = true;
 
-// current state is finished when the next state is free (waiting)
-async function validateState(id) {
     while (isRunning) {
-        await new Promise(resolve => setTimeout(resolve, 500));
-        console.log(getShuttles());
-        const shuttle = getShuttles().find(s => s.id === id);
+        await new Promise(resolve => setTimeout(resolve, 1000));
+
+        const shuttle = await (lastEvtType === null ? findShuttle(id, 3) : findShuttle(id));
+
         if (shuttle.next_state === NEXT_STATES[0]) {
+            if (checkLast && lastEvtType !== null) {
+                if ((shuttle.state !== lastEvtType
+                    && !(shuttle.state === NEXT_STATES[0] && lastEvtType === NEXT_STATES[3]))) {
+                    if (checkCount < checkLimit) {
+                        checkCount++;
+                        continue;
+                    }
+                    throw new Error(`Shuttle ${id} expected state ${lastEvtType}, got ${shuttle.state}`);
+                }
+            }
             return;
         }
     }
 }
 
-// TODO: use event driven updates instead of polling with getShuttles
-export function setGetShuttles(func) {
-    getShuttles = func;
+// allow multiple tries to find shuttles, useful for processing the first event
+async function findShuttle(id, tries = 1) {
+    for (let i = 0; i < tries; i++) {
+        const shuttle = getShuttles().find(s => s.id === id);
+        if (shuttle !== undefined) return shuttle;
+        if (i < tries - 1) await new Promise(resolve => setTimeout(resolve, 1000));
+    }
+    throw new Error(`Couldn't find shuttle with id ${id}`);
 }
 
-// TODO: fix stopTest to also stop in-execution promises
+// in-execution promises will cancel their tasks and resolve silently
 export function stopTest() {
     isRunning = false;
+}
+
+export function setGetShuttles(func) {
+    getShuttles = func;
 }
 
 // TODO: replace console errors with UI alerts in App.jsx
@@ -101,13 +134,15 @@ function warnUser(e) {
 
 // api call wrappers
 function addShuttle() {
+    if (!isRunning) return;
     return fetch("/api/shuttles", {
         method: "POST",
         headers: { "Content-Type": "application/json" },
     }).then(assertOk);
 }
 
-function setState(shuttleId, state) {
+function setNextState(shuttleId, state) {
+    if (!isRunning) return;
     return fetch(`/api/shuttles/${shuttleId}/set-next-state`, {
         method: "POST",
         headers: { "Content-Type": "application/json" },
@@ -115,11 +150,14 @@ function setState(shuttleId, state) {
     }).then(assertOk);
 }
 
+/*
 function clearData() {
+    if (!isRunning) return;
     return fetch(`/api/events/today?keepShuttles=false`, {
         method: "DELETE",
     }).then(assertOk);
 }
+*/
 
 // helper to reveal http errors
 async function assertOk(res) {
